requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.4.0=py310h06a4308_0
aiohttp=3.8.5=py310h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py310h06a4308_0
asynctest=0.13.0=py_0
attrs=23.1.0=py310h06a4308_0
blas=1.0=mkl
blinker=1.6.2=py310h06a4308_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py310h6a678d5_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py310h06a4308_0
contourpy=1.0.5=py310hdb19cb5_0
cryptography=41.0.3=py310h130f0dd_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h9c0eb46_1
daal4py=2023.1.1=py310h3c18c91_0
dal=2023.1.1=hdb19cb5_48680
dbus=1.13.18=hb2f20db_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
flatbuffers=2.0.0=h2531618_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py310h5eee18b_0
future=0.18.3=py310h06a4308_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.48.2=h5bf31a4_0
grpcio=1.48.2=py310h5bf31a4_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.9.0=py310he06866b_0
hdf5=1.12.1=h70be1eb_2
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
importlib-metadata=6.0.0=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h5eee18b_1
keras=2.12.0=py310h06a4308_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.20.1=h568e23c_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=ha637b67_0
libcurl=8.2.1=h91b91d3_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.52.0=ha637b67_1
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.44.2=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py310h06a4308_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib=3.8.0=py310h06a4308_0
matplotlib-base=3.8.0=py310h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpi=1.0=mpich
mpich=4.1.1=hbae89fd_0
multidict=6.0.2=py310h5eee18b_0
munkres=1.1.4=py_0
mysql=5.7.24=he378463_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.7=py310h85018f9_0
numpy=1.23.5=py310h5f9d8c6_1
numpy-base=1.23.5=py310hb5e798b_1
oauthlib=3.2.2=py310h06a4308_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.1=py310h06a4308_0
pandas=2.1.1=py310h1128e8f_0
pcre=8.45=h295c915_0
pillow=10.0.1=py310ha6cbd5a_0
pip=23.3=py310h06a4308_0
plotly=5.9.0=py310h06a4308_0
ply=3.11=py310h06a4308_0
protobuf=3.20.3=py310h6a678d5_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py310h06a4308_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.10=py310h6a678d5_0
pyqt5-sip=12.13.0=py310h5eee18b_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h7a1cb2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=1.13.1=cpu_py310ha02dd7b_1
pytorch-mutex=1.0=cpu
pytz=2023.3.post1=py310h06a4308_0
pyyaml=6.0.1=py310h5eee18b_0
qt-main=5.15.2=h5b8104b_9
qt-webengine=5.15.9=h9ab4d14_7
qtwebkit=5.212=h3fafdc1_5
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.31.0=py310h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.3.0=py310h1128e8f_0
scikit-learn-intelex=2023.1.1=py310h06a4308_0
scipy=1.11.3=py310h5f9d8c6_0
seaborn=0.12.2=py310h06a4308_0
setuptools=68.0.0=py310h06a4308_0
sip=6.7.12=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.2=py310h06a4308_0
tensorboard=2.12.1=py310h06a4308_0
tensorboard-data-server=0.7.0=py310h52d8a92_0
tensorboard-plugin-wit=1.8.1=py310h06a4308_0
tensorflow=2.12.0=mkl_py310hc7ea715_0
tensorflow-base=2.12.0=mkl_py310he5f8e37_0
tensorflow-estimator=2.12.0=py310h06a4308_0
termcolor=2.1.0=py310h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py310h06a4308_0
torchaudio=0.13.1=py310_cpu
torchvision=0.15.2=cpu_py310h83e0c9b_0
tornado=6.3.3=py310h5eee18b_0
typing-extensions=4.7.1=py310h06a4308_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py310h06a4308_0
werkzeug=2.2.3=py310h06a4308_0
wheel=0.41.2=py310h06a4308_0
wrapt=1.14.1=py310h5eee18b_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py310h5eee18b_0
zipp=3.11.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
